refactor init sagas and url builder
