+++ app/client/src/entities/Engine/AppEditorEngine.ts
@@ -1,0 +1,159 @@
+import {
+  fetchAppThemesAction,
+  fetchSelectedAppThemeAction,
+} from "actions/appThemingActions";
+import {
+  fetchDatasources,
+  fetchMockDatasources,
+} from "actions/datasourceActions";
+import {
+  fetchGitStatusInit,
+  remoteUrlInputValue,
+  resetPullMergeStatus,
+} from "actions/gitSyncActions";
+import { restoreRecentEntitiesRequest } from "actions/globalSearchActions";
+import { resetEditorSuccess } from "actions/initActions";
+import { fetchJSCollections } from "actions/jsActionActions";
+import {
+  fetchAllPageEntityCompletion,
+  fetchPage,
+  fetchPageDSLs,
+  fetchPageSuccess,
+} from "actions/pageActions";
+import {
+  executePageLoadActions,
+  fetchActions,
+} from "actions/pluginActionActions";
+import { fetchPluginFormConfigs, fetchPlugins } from "actions/pluginActions";
+import {
+  ApplicationPayload,
+  ReduxActionErrorTypes,
+  ReduxActionTypes,
+} from "ce/constants/ReduxActionConstants";
+import { addBranchParam } from "constants/routes";
+import { APP_MODE } from "entities/App";
+import { call, put, select } from "redux-saga/effects";
+import { failFastApiCalls } from "sagas/InitSagas";
+import { getCurrentApplication } from "selectors/editorSelectors";
+import { getCurrentGitBranch } from "selectors/gitSyncSelectors";
+import AnalyticsUtil from "utils/AnalyticsUtil";
+import history from "utils/history";
+import PerformanceTracker, {
+  PerformanceTransactionName,
+} from "utils/PerformanceTracker";
+import AppEngine, { AppEnginePayload } from ".";
+
+export default class AppEditorEngine extends AppEngine {
+  constructor(mode: APP_MODE) {
+    super(mode);
+  }
+
+  /**
+   * this saga is called once then application is loaded.
+   * It will hold the editor in uninitialized till all the apis/actions are completed
+   *
+   * @param AppEnginePayload
+   * @returns
+   */
+  *bootstrapEditor(payload: AppEnginePayload): any {
+    yield call(super.bootstrapEditor.bind(this), payload);
+    yield put(resetEditorSuccess());
+  }
+
+  startPerformanceTracking() {
+    PerformanceTracker.startAsyncTracking(
+      PerformanceTransactionName.INIT_EDIT_APP,
+    );
+  }
+
+  stopPerformanceTracking() {
+    PerformanceTracker.stopAsyncTracking(
+      PerformanceTransactionName.INIT_EDIT_APP,
+    );
+  }
+
+  *loadAppEntities(toLoadPageId: string, applicationId: string): any {
+    const initActionsCalls = [
+      fetchPage(toLoadPageId, true),
+      fetchActions({ applicationId }, []),
+      fetchJSCollections({ applicationId }),
+      fetchSelectedAppThemeAction(applicationId),
+      fetchAppThemesAction(applicationId),
+      fetchPlugins(),
+      fetchDatasources(),
+      fetchMockDatasources(),
+      fetchPageDSLs(),
+    ];
+
+    const successActionEffects = [
+      ReduxActionTypes.FETCH_JS_ACTIONS_SUCCESS,
+      ReduxActionTypes.FETCH_ACTIONS_SUCCESS,
+      ReduxActionTypes.FETCH_APP_THEMES_SUCCESS,
+      ReduxActionTypes.FETCH_SELECTED_APP_THEME_SUCCESS,
+      fetchPageSuccess().type,
+      ReduxActionTypes.FETCH_PLUGINS_SUCCESS,
+      ReduxActionTypes.FETCH_DATASOURCES_SUCCESS,
+      ReduxActionTypes.FETCH_MOCK_DATASOURCES_SUCCESS,
+    ];
+    const failureActionEffects = [
+      ReduxActionErrorTypes.FETCH_JS_ACTIONS_ERROR,
+      ReduxActionErrorTypes.FETCH_ACTIONS_ERROR,
+      ReduxActionErrorTypes.FETCH_APP_THEMES_ERROR,
+      ReduxActionErrorTypes.FETCH_SELECTED_APP_THEME_ERROR,
+      ReduxActionErrorTypes.FETCH_PAGE_ERROR,
+      ReduxActionErrorTypes.FETCH_PLUGINS_ERROR,
+      ReduxActionErrorTypes.FETCH_DATASOURCES_ERROR,
+      ReduxActionErrorTypes.FETCH_MOCK_DATASOURCES_ERROR,
+    ];
+    const allActionCalls: boolean = yield failFastApiCalls(
+      initActionsCalls,
+      successActionEffects,
+      failureActionEffects,
+    );
+
+    if (!allActionCalls) return;
+
+    const pluginFormCall: boolean = yield failFastApiCalls(
+      [fetchPluginFormConfigs()],
+      [ReduxActionTypes.FETCH_PLUGIN_FORM_CONFIGS_SUCCESS],
+      [ReduxActionErrorTypes.FETCH_PLUGIN_FORM_CONFIGS_ERROR],
+    );
+    if (!pluginFormCall) return;
+
+    yield put({
+      type: ReduxActionTypes.FETCH_PLUGIN_AND_JS_ACTIONS_SUCCESS,
+    });
+    yield put(fetchAllPageEntityCompletion([executePageLoadActions()]));
+  }
+
+  public *completeChore() {
+    const currentApplication: ApplicationPayload = yield select(
+      getCurrentApplication,
+    );
+    AnalyticsUtil.logEvent("EDITOR_OPEN", {
+      appId: currentApplication.id,
+      appName: currentApplication.name,
+    });
+    yield put({
+      type: ReduxActionTypes.INITIALIZE_EDITOR_SUCCESS,
+    });
+  }
+
+  public *initiateGit(applicationId: string) {
+    const branchInStore: string = yield select(getCurrentGitBranch);
+    yield put(
+      restoreRecentEntitiesRequest({
+        applicationId,
+        branch: branchInStore,
+      }),
+    );
+    // init of temporay remote url from old application
+    yield put(remoteUrlInputValue({ tempRemoteUrl: "" }));
+    // add branch query to path and fetch status
+    if (branchInStore) {
+      history.replace(addBranchParam(branchInStore));
+      yield put(fetchGitStatusInit());
+    }
+    yield put(resetPullMergeStatus());
+  }
+}
