+++ app/client/src/entities/Engine/factory.ts
@@ -1,0 +1,9 @@
+import { APP_MODE } from "entities/App";
+import AppEngine from ".";
+import AppEditorEngine from "./AppEditorEngine";
+import AppViewerEngine from "./AppViewerEngine";
+
+export function AppEngineFactory(mode: APP_MODE): AppEngine {
+  if (mode === APP_MODE.EDIT) return new AppEditorEngine(mode);
+  return new AppViewerEngine(mode);
+}
