+++ app/client/src/sagas/PageSagas.tsx
@@ -920,9 +920,6 @@ function* fetchPageDSLSaga(pageId: string) {
 
 export function* populatePageDSLsSaga() {
   try {
-    yield put({
-      type: ReduxActionTypes.POPULATE_PAGEDSLS_INIT,
-    });
     const pageIds: string[] = yield select((state: AppState) =>
       state.entities.pageList.pages.map((page: Page) => page.pageId),
     );
@@ -1080,5 +1077,6 @@ export default function* pageSagas() {
       generateTemplatePageSaga,
     ),
     takeLatest(ReduxActionTypes.SET_PAGE_ORDER_INIT, setPageOrderSaga),
+    takeLatest(ReduxActionTypes.POPULATE_PAGEDSLS_INIT, populatePageDSLsSaga),
   ]);
 }
